✨ [feat] #47 화면 분리 플로우 설계
- 영상 병합 후 결과를 확인할 수 있는 프로젝트 미리보기 화면 및 네비게이션 케이스 추가
- VideoManager가 병합된 영상을 직접 사진 앨범에 저장하지 않고 URL만 반환하도록 리팩토링
- 사진 앨범 저장 로직을 PhotoLibrarySaver라는 새 클래스로 분리
- 변경된 흐름에 맞게 관련 뷰모델 및 네비게이션 로직 업데이트
