+++ Chalkak/Common/Util/Enum/Path.swift
@@ -11,4 +11,5 @@ enum Path: Hashable {
     case clipEdit(clipURL: URL, isFirstShoot: Bool, guide: Guide?)
     case overlay(clipID: String)
     case boundingBox(guide: Guide, isFirstShoot: Bool)
+    case projectPreview(finalVideoURL: URL)
 }
